move solana helpers
